Implement proof-of-concept recursive exploration of nodes
Use the Roslyn mechanism of "Find all references" for that. Enable to work with methods, constructors and property accessors. Add buttons to explicitly show certain node in the code and inspect it (find its children).

Fix the text highlighting and text selecting problems after opening a new document - it no longer has a corresponding snapshot, because a new solution is created. Therefore, we need to obtain the current document using the former's ID. Highlighting mechanism must enable to highlight text in a delayed manner, because the tagger is loaded after the code asking it to highlight already ended; thus missing the highlight event invoke.